Validação de excluir hospede caso tenha reserva ativa

+++ src/app/pages/guests/guests.component.ts
@@ -5,6 +5,9 @@ import { ModalComponent } from '../../components/modal/modal.component';
 import { FormGuestComponent } from '../../components/form-guest/form-guest.component';
 import { AlertComponent } from '../../components/alert/alert.component';
 import { CommomButtonComponent } from '../../components/commom-button/commom-button.component';
+import { ReservationsService } from '../../services/reservations.service';
+import { TReservations } from '../../@types/reservations';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-guests',
@@ -13,20 +16,29 @@ import { CommomButtonComponent } from '../../components/commom-button/commom-but
     FormGuestComponent,
     AlertComponent,
     CommomButtonComponent,
+    CommonModule,
   ],
   templateUrl: './guests.component.html',
   styleUrl: './guests.component.scss',
 })
 export class GuestsComponent implements OnInit {
   guests: Array<TGuests> = [];
+  reservation!: TReservations;
+  reservations: Array<TReservations> = [];
+  guestHasReservation: boolean = false;
   guestId: string = '';
+  guestIdHasReservation: string = '';
   viewOrNo: boolean = false;
 
-  constructor(private guestService: GuestsService) {}
+  constructor(
+    private guestService: GuestsService,
+    private reservationService: ReservationsService
+  ) {}
 
   ngOnInit(): void {
     this.loadFirstGuest();
     this.loadGuests();
+    this.getReservations();
   }
 
   loadFirstGuest(): void {
@@ -44,6 +56,9 @@ export class GuestsComponent implements OnInit {
     this.guestService.getGuests().subscribe({
       next: (guests) => {
         this.guests = guests;
+        // if (guests.length > 0) {
+        //   this.loadGuestInformations(guests[0].id);
+        // }
       },
       error: (err) => console.error(err),
     });
@@ -58,16 +73,45 @@ export class GuestsComponent implements OnInit {
     });
   }
 
-  removeGuest(id: string): void {
-    this.guestService.deleteGuest(id).subscribe({
-      next: () => {
-        this.loadGuests();
+  getReservations(): void {
+    this.reservationService.getReservations().subscribe({
+      next: (reserv) => {
+        this.reservations = reserv;
+        console.log('EIS AQUI AS RESERVA:', this.reservations);
       },
-      error: (err) => console.error(err),
     });
   }
 
+  validationGustHasReservation(id: string): boolean {
+    this.reservations.forEach((element) => {
+      if (element.guestId == id) {
+        this.guestIdHasReservation = element.guestId;
+        console.log('GUEST POSSUI SIM RESERVA', element.guestId);
+      }
+    });
+    this.guestIdHasReservation == id
+      ? (this.guestHasReservation = true)
+      : (this.guestHasReservation = false);
+    return true;
+  }
+
+  removeGuest(id: string): void {
+    if (this.guestIdHasReservation != id) {
+      this.guestHasReservation = false;
+      this.guestService.deleteGuest(id).subscribe({
+        next: (value) => {
+          console.log('GUEST EXCLUIDO!', value);
+          this.loadGuests();
+        },
+        error: (err) => console.error(err),
+      });
+    }
+  }
+
   openModal(id: string, modalId: string) {
+    if (modalId == 'modalDeleteGuest') {
+      this.validationGustHasReservation(id);
+    }
     this.guestId = id;
     this.toggleModal(modalId, true);
   }
